Move to S3-hosted static images with img
This commit removes a large amount of previously-built code
using the GatsbyImage and StaticImage classes in order to use the
standard html img.

It also removes image files from the git, in order to use URLs to
images hosted statically on AWS S3.

A style sheet is also added to control the size of the image on
the landing page.