Refactor internal file system to store blog and portfolio mdx files
This commit restructures the internal filesystem to allow the storage
of blog and portfolio mdx files in a neat and logical manner. It also
adjusts the query used by the blog index to match, and adds exemplar
code to the portfolio index that is able to load its own relevant
files.