Add portfolio directory with index and dynamic file
This commit adds a dynamic file and index to a portfolio directory,
allowing portfolio/etc pages to be correctly indexed. Notably it
introduces a bug where the blog page 'example' can be indexed at
either blog/example or portfolio/example - this is a minor issue
but will need to be fixed.